+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 ### Fixed
 - Escape special characters in search queries
 - Handle colons in the body of search queries
+- Add deadlock handling in `ClusterProcess`
 - Scheduled tests-regression.yaml to run every Monday 12 AM EST
 - Renamed tests.yaml as "tests-unit.yaml"
+
 
 ## 2021-04-15 -- v0.5.6
 ### Added